Mutex fstrans #435
There are regions in the ZFS code where it is desirable to be able
to be set PF_FSTRANS while a specific mutex is held. The ZFS code
could be updated to set/clear this flag in all the correct places,
but this is undesirable for a few reasons.

1) It would require changes to a significant amount of the ZFS
code. This would complicate applying patches from upstream.

2) It would be easy to accidentally miss a critical region in
the initial patch or to have an future change introduce a
new one.

Both of these concerns can be addressed by adding a new mutex type
which is responsible for managing PF_FSTRANS. This lets us make a
a small change to the ZFS source where the mutex is initialized
and then be certain that all future use of that mutex will be safe.

NOTES: The ht_locks are no longer aligned on 64-byte boundaries.
We've never studied if this is actually critical for performance
when there are a large number of hash buckets. The dbuf hash has
never made this optimization.

Signed-off-by: Brian Behlendorf behlendorf1@llnl.gov
Issue #3050
Issue #3055